requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv==0.19.2
yt-dlp>=2023.11.16
aiofiles>=23.2.0
aiosqlite>=0.18.0
